refactor: make services package
