+++ src/main/java/org/team1540/robot2025/services/AlertManager.java
@@ -1,4 +1,4 @@
-package org.team1540.robot2025;
+package org.team1540.robot2025.services;
 
 import com.ctre.phoenix6.CANBus;
 import edu.wpi.first.hal.can.CANStatus;
@@ -6,6 +6,7 @@ import edu.wpi.first.math.filter.Debouncer;
 import edu.wpi.first.wpilibj.Alert;
 import edu.wpi.first.wpilibj.DriverStation;
 import edu.wpi.first.wpilibj.RobotController;
+import org.team1540.robot2025.Constants;
 import org.team1540.robot2025.generated.TunerConstants;
 
 /**
