+++ src/main/java/org/team1540/robot2025/services/MechanismVisualizer.java
@@ -1,4 +1,4 @@
-package org.team1540.robot2025;
+package org.team1540.robot2025.services;
 
 import edu.wpi.first.math.geometry.Pose3d;
 import edu.wpi.first.math.geometry.Rotation2d;
