+++ src/main/java/org/team1540/robot2025/subsystems/intake/CoralIntake.java
@@ -12,7 +12,7 @@ import edu.wpi.first.wpilibj2.command.SubsystemBase;
 import org.littletonrobotics.junction.AutoLogOutput;
 import org.littletonrobotics.junction.Logger;
 import org.team1540.robot2025.Constants;
-import org.team1540.robot2025.MechanismVisualizer;
+import org.team1540.robot2025.services.MechanismVisualizer;
 import org.team1540.robot2025.util.LoggedTunableNumber;
 
 public class CoralIntake extends SubsystemBase {
